% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Label data based on a code book.}
\usage{
codebook(
  data,
  path = "dictionary.yml",
  label = FALSE,
  to_labelled = FALSE,
  .include = NULL,
  .exclude = NULL
)
}
\arguments{
\item{data}{The data to label.}

\item{path}{The path to YAML code book.}

\item{label}{Should column names also be renamed according to code book labels?}

\item{to_labelled}{Should the code book be applied using the labelled package?}

\item{.include}{An optional character vector of column names to
apply the code book to.}

\item{.exclude}{An optional character vector of column names to not apply
the code book to. Ignored if \code{.include} is specified.}
}
\value{
An updated copy of \code{data}.
}
\description{
Label data based on a code book.
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)
codebook(ex, system.file("codebook.yml", package = "codebook"))
}
