% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreak-package.R
\docType{package}
\name{codebreak-package}
\alias{codebreak}
\alias{codebreak-package}
\title{codebreak: Label Data Using a YAML Codebook}
\description{
A light-weight framework for labeling coded data using a codebook saved as YAML text file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nt-williams/codebreak}
  \item Report bugs at \url{https://github.com/nt-williams/codebreak/issues}
}

}
\author{
\strong{Maintainer}: Nick Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID})

}
\keyword{internal}
