% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Label data based on a data dictionary}
\usage{
dictionary(.data, path = "dictionary.json")
}
\arguments{
\item{.data}{The data to label.}

\item{path}{The path to JSON data dictionary.}
}
\value{
The data with the columns found in the data dictionary labeled.
}
\description{
Label data based on a data dictionary
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5)
)
dictionary(ex, system.file("dictionary.json", package = "dictionary"))
}
