% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Rename columns using code book labels}
\usage{
label(data, dict, .path = NULL, .include = NULL, .exclude = NULL)
}
\arguments{
\item{data}{The data to label.}

\item{.include}{An optional character vector of column names to
apply the code book to.}

\item{.exclude}{An optional character vector of column names to not apply
the code book to. Ignored if \code{.include} is specified.}

\item{path}{The path to YAML code book.}
}
\value{
The data with the columns found in the code book renamed.
}
\description{
Rename columns using code book labels
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)
label(ex, system.file("codebook.yml", package = "codebook"))
}
