% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{Codebook}
\alias{Codebook}
\title{R6 Class representing a codebook}
\description{
Decode and label data using a codebook saved as a YAML text file.
}
\examples{

## ------------------------------------------------
## Method `Codebook$new`
## ------------------------------------------------

Codebook$new(system.file("codebook.yml", package = "codebreak"))

## ------------------------------------------------
## Method `Codebook$decode`
## ------------------------------------------------

ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)

cb <- Codebook$new(system.file("codebook.yml", package = "codebreak"))
cb$decode(ex)

## ------------------------------------------------
## Method `Codebook$label`
## ------------------------------------------------

ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)
cb <- Codebook$new(system.file("codebook.yml", package = "codebreak"))
cb$label(ex)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to a YAML text file.}

\item{\code{codebook}}{The codebook.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Codebook-new}{\code{Codebook$new()}}
\item \href{#method-Codebook-decode}{\code{Codebook$decode()}}
\item \href{#method-Codebook-label}{\code{Codebook$label()}}
\item \href{#method-Codebook-print}{\code{Codebook$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codebook-new"></a>}}
\if{latex}{\out{\hypertarget{method-Codebook-new}{}}}
\subsection{Method \code{new()}}{
Import a YAML codebook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codebook$new(path = "codebook.yml")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to the YAML codebook.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Codebook` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Codebook$new(system.file("codebook.yml", package = "codebreak"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codebook-decode"></a>}}
\if{latex}{\out{\hypertarget{method-Codebook-decode}{}}}
\subsection{Method \code{decode()}}{
Decode data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codebook$decode(
  data,
  label = FALSE,
  as_labelled = FALSE,
  .include = NULL,
  .exclude = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to decode.}

\item{\code{label}}{Should column names also be renamed according to codebook labels?}

\item{\code{as_labelled}}{Should the codebook be applied using the labelled package?}

\item{\code{.include}}{An optional character vector of column names to
apply the codebook to.}

\item{\code{.exclude}}{An optional character vector of column names to not apply
the codebook to. Ignored if \code{.include} is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An updated copy of \code{data}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)

cb <- Codebook$new(system.file("codebook.yml", package = "codebreak"))
cb$decode(ex)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codebook-label"></a>}}
\if{latex}{\out{\hypertarget{method-Codebook-label}{}}}
\subsection{Method \code{label()}}{
Label data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codebook$label(data, as_labelled = FALSE, .include = NULL, .exclude = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to label.}

\item{\code{as_labelled}}{Should the codebook labels be applied using the labelled package?}

\item{\code{.include}}{An optional character vector of column names to
apply the codebook to.}

\item{\code{.exclude}}{An optional character vector of column names to not apply
the codebook to. Ignored if \code{.include} is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)
cb <- Codebook$new(system.file("codebook.yml", package = "codebreak"))
cb$label(ex)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codebook-print"></a>}}
\if{latex}{\out{\hypertarget{method-Codebook-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codebook$print()}\if{html}{\out{</div>}}
}

}
}
