% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Label data based on a code book.}
\usage{
codebook(.data, path = "dictionary.json")
}
\arguments{
\item{.data}{The data to label.}

\item{path}{The path to JSON code book.}
}
\value{
The data with the columns found in the code book labeled.
}
\description{
Label data based on a code book.
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5)
)
codebook(ex, system.file("codebook.json", package = "codebook"))
}
