% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Label data based on a codebook.}
\usage{
codebook(
  data,
  path = "codebook.yml",
  label = FALSE,
  as_labelled = FALSE,
  .include = NULL,
  .exclude = NULL
)
}
\arguments{
\item{data}{The data to label.}

\item{path}{The path to YAML codebook.}

\item{label}{Should column names also be renamed according to codebook labels?}

\item{as_labelled}{Should the codebook be applied using the labelled package?}

\item{.include}{An optional character vector of column names to
apply the codebook to.}

\item{.exclude}{An optional character vector of column names to not apply
the codebook to. Ignored if \code{.include} is specified.}
}
\value{
An updated copy of \code{data}.
}
\description{
Label data based on a codebook.
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)

codebook(ex, system.file("codebook.yml", package = "codebook"))
}
