% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{label}
\alias{label}
\title{Rename columns using code book labels}
\usage{
label(
  data,
  path = "codebook.yml",
  as_labelled = FALSE,
  .include = NULL,
  .exclude = NULL
)
}
\arguments{
\item{data}{The data to label.}

\item{path}{The path to YAML codebook.}

\item{as_labelled}{Should the codebook labels be applied using the labelled package?}

\item{.include}{An optional character vector of column names to
apply the codebook to.}

\item{.exclude}{An optional character vector of column names to not apply
the codebook to. Ignored if \code{.include} is specified.}
}
\value{
The data with the columns found in the code book renamed.
}
\description{
Rename columns using code book labels
}
\examples{
ex <- data.frame(
  x = c(1, 2, 5, 3, 4),
  y = c(0, 1, 1, 0, 1),
  z = rnorm(5),
  w = c(1, 1, 0, 1, 1)
)

label(ex, system.file("codebook.yml", package = "codebook"))
}
